ignore virtual interfaces
Getting a fix for this upstream has been difficult; the upstream
position is that it's better for hsflowd to send all counters and for
sflow aggregators to do their own filtering. The merits of this position
are debatable but certainly not outright wrong.

In our case, getting ganglia to do the filtering is either difficult or
impossible, so we hack hsflowd. This local patch to disable virtual
interfaces probably disables metrics for VMs, etc.--functionality we do
not use.

Our previous approach was to completely ignore virtual adaptors; this
fails now that we are assigning IPs to virtual adapters (team0, etc.).
The new approach detects virtual adaptors and refrains from updating
counters on them.